+++ viewlegacy/src/androidMain/kotlin/tech/skot/core/components/SKListViewProxy.kt
@@ -12,6 +12,7 @@ open class SKListViewProxy(
     private val reverse: Boolean = false,
     private val animate: Boolean = true,
     private val animateItem: Boolean = false,
+    private val infiniteScroll: Boolean = false,
 ) : SKComponentViewProxy<RecyclerView>(), SKListVC {
     private val itemsLD: MutableSKLiveData<List<Triple<SKComponentViewProxy<*>, Any, (() -> Unit)?>>> = MutableSKLiveData(emptyList())
 
@@ -46,7 +47,7 @@ open class SKListViewProxy(
         activity: SKActivity,
         fragment: Fragment?,
         binding: RecyclerView,
-    ) = SKListView(layoutMode, reverse, animate, animateItem, this, activity, fragment, binding).apply {
+    ) = SKListView(layoutMode, reverse, animate, animateItem, infiniteScroll, this, activity, fragment, binding).apply {
         itemsLD.observe {
             onItems(it)
         }