+++ viewmodelTests/src/main/kotlin/tech/skot/core/di/CoreViewInjectorMock.kt
@@ -1,8 +1,46 @@
 package tech.skot.core.di
 
-import tech.skot.core.components.*
-import tech.skot.core.components.inputs.*
-import tech.skot.core.components.presented.*
+import tech.skot.core.components.SKBoxVC
+import tech.skot.core.components.SKBoxViewMock
+import tech.skot.core.components.SKComponentVC
+import tech.skot.core.components.SKFrameVC
+import tech.skot.core.components.SKFrameViewMock
+import tech.skot.core.components.SKListVC
+import tech.skot.core.components.SKListViewMock
+import tech.skot.core.components.SKLoaderVC
+import tech.skot.core.components.SKLoaderViewMock
+import tech.skot.core.components.SKPagerVC
+import tech.skot.core.components.SKPagerViewMock
+import tech.skot.core.components.SKPagerWithTabsVC
+import tech.skot.core.components.SKPagerWithTabsViewMock
+import tech.skot.core.components.SKRootStackViewMock
+import tech.skot.core.components.SKScreenVC
+import tech.skot.core.components.SKStackVC
+import tech.skot.core.components.SKStackViewMock
+import tech.skot.core.components.SKWebViewVC
+import tech.skot.core.components.SKWebViewViewMock
+import tech.skot.core.components.inputs.SKButtonVC
+import tech.skot.core.components.inputs.SKButtonViewMock
+import tech.skot.core.components.inputs.SKComboVC
+import tech.skot.core.components.inputs.SKComboViewMock
+import tech.skot.core.components.inputs.SKImageButtonVC
+import tech.skot.core.components.inputs.SKImageButtonViewMock
+import tech.skot.core.components.inputs.SKInputVC
+import tech.skot.core.components.inputs.SKInputViewMock
+import tech.skot.core.components.inputs.SKInputWithSuggestionsVC
+import tech.skot.core.components.inputs.SKInputWithSuggestionsViewMock
+import tech.skot.core.components.inputs.SKSimpleInputVC
+import tech.skot.core.components.inputs.SKSimpleInputViewMock
+import tech.skot.core.components.presented.SKAlertVC
+import tech.skot.core.components.presented.SKAlertViewMock
+import tech.skot.core.components.presented.SKBottomSheetVC
+import tech.skot.core.components.presented.SKBottomSheetViewMock
+import tech.skot.core.components.presented.SKDialogVC
+import tech.skot.core.components.presented.SKDialogViewMock
+import tech.skot.core.components.presented.SKSnackBarVC
+import tech.skot.core.components.presented.SKSnackBarViewMock
+import tech.skot.core.components.presented.SKWindowPopupVC
+import tech.skot.core.components.presented.SKWindowPopupViewMock
 import tech.skot.core.view.Icon
 
 class CoreViewInjectorMock : CoreViewInjector {
@@ -57,6 +95,7 @@ class CoreViewInjectorMock : CoreViewInjector {
         reverse: Boolean,
         animate: Boolean,
         animateItem: Boolean,
+        infiniteScroll: Boolean
     ): SKListVC {
         return SKListViewMock(layoutMode, reverse, animate, animateItem)
     }